backend/requirements/base.txt
# Django
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.0
django-environ==0.11.2
django-filter==23.4
django-extensions==3.2.3
django-nested-admin==4.0.2
django-storages==1.14.2
Pillow>=10.2.0

# API Documentation
drf-yasg>=1.21.7

# Authentication
djangorestframework-simplejwt==5.3.0
django-allauth==0.58.2
# django-rest-auth==0.9.5  # Loại bỏ, đã bị deprecated
dj-rest-auth==5.0.1
google-auth==2.23.4

# Database
psycopg2-binary==2.9.9
dj-database-url==2.1.0

# Blockchain - sử dụng phiên bản tương thích với Python 3.11
web3>=6.11.0
py-solc-x>=1.1.1
# eth-brownie==1.19.3 - Removed due to Cython compilation issues
# py-evm==0.7.0a1 - Removed due to dependencies on cytoolz

# Testing and development
pytest==7.4.3
pytest-django==4.7.0
pytest-cov==4.1.0
factory-boy==3.3.0
faker==20.1.0
black==22.10.0
isort==5.12.0
flake8==6.1.0

# Deployment
gunicorn==21.2.0
whitenoise==6.6.0

# Utilities
python-dateutil==2.8.2
pyyaml==6.0.1
requests==2.31.0

# Thêm setuptools mới
setuptools>=69.0.0
